# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EasyEscrow.ai API
  version: 1.0.0
  description: Non-custodial atomic swap platform for NFTs and SOL on Solana, plus institutional cross-border USDC escrow payments. Browse the Documentation section for guides, or jump to API Reference for endpoint details.
  contact:
    name: EasyEscrow.ai Support
    email: support@easyescrow.ai
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://staging-api.easyescrow.ai
    description: Staging environment (Devnet)
  - url: https://api.easyescrow.ai
    description: Production (Mainnet)
...
paths:
...
  /api/v1/institution/clients:
    get:
      tags:
        - Institution Clients
      summary: List clients
      description: List institution clients (paginated). Returns public profile fields only.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
        - in: query
          name: includeArchived
          schema:
            type: boolean
            default: false
          description: Include archived clients
      responses:
        '200':
          description: Client list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InstitutionPublicProfile'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
  /api/v1/institution/clients/search:
    get:
      tags:
        - Institution Clients
      summary: Search clients
      description: Search clients by name, industry, or country. Returns public profile fields only.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search by company name, legal name, or trading name
        - in: query
          name: industry
          schema:
            type: string
          description: Filter by industry
        - in: query
          name: country
          schema:
            type: string
          description: Filter by country
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InstitutionPublicProfile'
                  count:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
  /api/v1/institution/clients/{id}/profile:
    get:
      tags:
        - Institution Clients
      summary: Get client public profile
      description: Retrieve non-sensitive public profile fields for a specific client.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Client ID
      responses:
        '200':
          description: Public profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/InstitutionPublicProfile'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/institution/clients/{id}/archive:
    put:
      tags:
        - Institution Clients
      summary: Archive/unarchive client
      description: Set or clear the archived flag on a client.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Client ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                archive:
                  type: boolean
                  description: Set to true to archive, false to unarchive
                  default: true
      responses:
        '200':
          description: Archive status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      companyName:
                        type: string
                      isArchived:
                        type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Wallet Management (new model) ──────────────────────────────────
...
  /api/v1/ai/analyze-escrow-doc/{escrow_id}:
    post:
      tags:
        - AI Analysis
      summary: Analyze escrow document (primary)
      description: |
        Trigger AI-powered analysis on an uploaded document associated with an escrow.
        The analysis evaluates document authenticity, risk indicators, and compliance factors.
        Client details (company name, address, industry) are included in the analysis context.

        **Rate Limited**: 10 requests per minute.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: escrow_id
          required: true
          schema:
            type: string
            format: uuid
          description: Escrow ID the document belongs to
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiAnalyzeRequest'
            example:
              fileId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              context:
                expectedAmount: 50000
                poNumber: "PO-2025-0042"
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AiAnalysisResult'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '503':
          description: AI analysis service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ai/escrow-doc-analysis/{escrow_id}:
    get:
      tags:
        - AI Analysis
      summary: Get escrow document analysis results (primary)
      description: Retrieve all AI analysis results for a specific escrow.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: escrow_id
          required: true
          schema:
            type: string
            format: uuid
          description: Escrow ID
      responses:
        '200':
          description: Analysis results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AiAnalysisResult'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Legacy AI analysis routes (backward compatibility aliases)
  /api/v1/ai/analyze/{escrow_id}:
    post:
      tags:
        - AI Analysis
      summary: Analyze document (legacy alias)
      description: |
        **Legacy route — use `/api/v1/ai/analyze-escrow-doc/{escrow_id}` instead.**

        Trigger AI-powered analysis on an uploaded document associated with an escrow.
        The analysis evaluates document authenticity, risk indicators, and compliance factors.

        **Rate Limited**: 10 requests per minute.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: escrow_id
          required: true
          schema:
            type: string
            format: uuid
          description: Escrow ID the document belongs to
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiAnalyzeRequest'
            example:
              fileId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              context:
                expectedAmount: 50000
                poNumber: "PO-2025-0042"
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AiAnalysisResult'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '503':
          description: AI analysis service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ai/analysis/{escrow_id}:
    get:
      tags:
        - AI Analysis
      summary: Get analysis results (legacy alias)
      description: |
        **Legacy route — use `/api/v1/ai/escrow-doc-analysis/{escrow_id}` instead.**

        Retrieve all AI analysis results for a specific escrow.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: escrow_id
          required: true
          schema:
            type: string
            format: uuid
          description: Escrow ID
      responses:
        '200':
          description: Analysis results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AiAnalysisResult'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Institution Admin ──────────────────────────────────────────────────
...
components:
...
  schemas:
...
    InstitutionPublicProfile:
      type: object
      description: Non-sensitive public profile for search/list results
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        legalName:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        jurisdiction:
          type: string
          nullable: true
        entityType:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        tier:
          type: string
          enum: [STANDARD, PREMIUM, ENTERPRISE]
        isRegulatedEntity:
          type: boolean
          nullable: true
        country:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AiAnalyzeRequest:
      type: object
      required:
        - fileId
      properties:
        fileId:
          type: string
          format: uuid
          description: ID of the uploaded file to analyze
        context:
          type: object
          description: Optional context to improve analysis accuracy
          properties:
            expectedAmount:
              type: number
              minimum: 0
              description: Expected payment amount for cross-referencing
              example: 50000
            poNumber:
              type: string
              maxLength: 100
              description: Purchase order number for verification
              example: "PO-2025-0042"
    AiAnalysisResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        escrowId:
          type: string
          format: uuid
        fileId:
          type: string
          format: uuid
        riskScore:
          type: number
          minimum: 0
          maximum: 100
          description: Risk score from 0 (low risk) to 100 (high risk)
          example: 15
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH]
          description: Categorized risk level
          example: "LOW"
        findings:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
                example: "DOCUMENT_AUTHENTICITY"
              severity:
                type: string
                enum: [INFO, WARNING, CRITICAL]
              message:
                type: string
                example: "Document appears authentic with consistent formatting"
          description: Individual analysis findings
        summary:
          type: string
          description: Human-readable summary of the analysis
          example: "Document passes all checks with low risk indicators"
        analyzedAt:
          type: string
          format: date-time

    # ── Institution Admin Schemas ───────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid wallet address format"
        timestamp:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for admin authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/v1/institution/auth/login
    SettlementAuthority:
      type: apiKey
      in: header
      name: X-Settlement-Authority-Key
      description: Settlement authority API key for release operations